+++ Samples~/SampleInjection/Scripts/Installers/ServicesAutoInstaller.cs
@@ -1,0 +1,15 @@
+﻿using AceLand.Injection;
+using AceLand.Sample.Injection.Scripts.Services;
+
+namespace AceLand.Sample.Injection.Scripts.Installers
+{
+    [AutoInstall]
+    public sealed class ServicesAutoInstaller : IGlobalInstaller
+    {
+        public void Install(IContainerBuilder builder)
+        {
+            builder.Register<IScoreService, ScoreService>(Lifetime.Singleton);
+            builder.RegisterEntryPoint<Helper>();
+        }
+    }
+}