+++ CHANGELOG.md
@@ -3,13 +3,21 @@
 All notable changes to this project will be documented in this file.
 
 --- 
+# Beta - published
+
+## [0.2.2] - 2026-08-14
+- [Graph] fixed open script fail on script is not in same filename
+- [Sample] added sample to learn 
 
 ## [0.2.1] - 2026-08-13
-- fix graph issues
+- fixed graph issues
 - add installer to graph
 
 ## [0.2.0] - 2026-08-13
 - beta published
+
+---
+# Dev - unpublished
 
 ## [0.1.1] - 2026-08-12
 dev optimize and bug fix